// graphql/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

/** A Field Group managed by ACF */
export type AcfFieldGroup = {
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Fields associated with an ACF Field Group */
export type AcfFieldGroupFields = {
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Connection between the CourseId_Fields type and the MediaItem type */
export type AcfMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'AcfMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type ContentNodeToContentTypeConnectionEdge = ContentTypeConnectionEdge & Edge & OneToOneConnection & {
  __typename?: 'ContentNodeToContentTypeConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: ContentType;
};
// ...
export type DefaultTemplate = ContentTemplate & {
  __typename?: 'DefaultTemplate';
  /** The name of the template */
  templateName?: Maybe<Scalars['String']['output']>;
};
// ...
export type Edge = {
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The connected node */
  node: Node;
};
// ...
export type LessonToLessonCategoryConnection = Connection & LessonCategoryConnection & {
  __typename?: 'LessonToLessonCategoryConnection';
  /** Edges for the LessonToLessonCategoryConnection connection */
  edges: Array<LessonToLessonCategoryConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<LessonCategory>;
  /** Information about pagination in a connection. */
  pageInfo: LessonToLessonCategoryConnectionPageInfo;
};
// ...
export type MediaDetails = {
  __typename?: 'MediaDetails';
  /** The filename of the mediaItem */
  file?: Maybe<Scalars['String']['output']>;
  /** The path to the mediaItem relative to the uploads directory */
  filePath?: Maybe<Scalars['String']['output']>;
  /** The height of the mediaItem */
  height?: Maybe<Scalars['Int']['output']>;
  /** Meta information associated with the mediaItem */
  meta?: Maybe<MediaItemMeta>;
  /** The available sizes of the mediaItem */
  sizes?: Maybe<Array<Maybe<MediaSize>>>;
  /** The width of the mediaItem */
  width?: Maybe<Scalars['Int']['output']>;
};


/** File details for a Media Item */
export type MediaDetailsSizesArgs = {
  exclude?: InputMaybe<Array<InputMaybe<MediaItemSizeEnum>>>;
  include?: InputMaybe<Array<InputMaybe<MediaItemSizeEnum>>>;
};

/** Represents uploaded media, including images, videos, documents, and audio files. */
export type MediaItem = ContentNode & DatabaseIdentifier & HierarchicalContentNode & HierarchicalNode & Node & NodeWithAuthor & NodeWithComments & NodeWithTemplate & NodeWithTitle & UniformResourceIdentifiable & {
// ...
  template?: Maybe<ContentTemplate>;
  /** The title of the post. This is currently just the raw title. An amendment to support rendered title needs to be made. */
  title?: Maybe<Scalars['String']['output']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']['output']>;
};
// ...
export type MediaItemConnectionEdge = {
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The connected mediaItem Node */
  node: MediaItem;
};
// ...
export type NodeWithFeaturedImageToMediaItemConnectionEdge = Edge & MediaItemConnectionEdge & OneToOneConnection & {
  __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge';
  /** Opaque reference to the nodes position in the connection. Value can be used with pagination args. */
  cursor?: Maybe<Scalars['String']['output']>;
  /** The node of the connection, without the edges */
  node: MediaItem;
};
// ...
export type Page = ContentNode & DatabaseIdentifier & HierarchicalContentNode & HierarchicalNode & MenuItemLinkable & Node & NodeWithAuthor & NodeWithComments & NodeWithContentEditor & NodeWithFeaturedImage & NodeWithPageAttributes & NodeWithRevisions & NodeWithTemplate & NodeWithTitle & Previewable & UniformResourceIdentifiable & WithAcfPageComponents & {
// ...
  pageComponents?: Maybe<PageComponents>;
// ...
};
// ...
export type PageComponentCloneFieldsBannerHomepageButton = AcfFieldGroup & AcfFieldGroupFields & PageComponentCloneFieldsBannerHomepageButton_Fields & {
  __typename?: 'PageComponentCloneFieldsBannerHomepageButton';
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageComponentCloneFieldsBannerHomepageButton&quot; Field Group */
  label?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageComponentCloneFieldsBannerHomepageButton&quot; Field Group */
  link?: Maybe<Scalars['String']['output']>;
};
// ...
export type PageComponents = AcfFieldGroup & AcfFieldGroupFields & PageComponents_Fields & {
  __typename?: 'PageComponents';
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;flexible_content&quot; Field Type added to the schema as part of the &quot;PageComponents&quot; Field Group */
  pageComponents?: Maybe<Array<Maybe<PageComponentsPageComponents_Layout>>>;
};

/** The &quot;PageComponentsPageComponentsBannerHomepage&quot; Field Group. Added to the Schema by &quot;WPGraphQL for ACF&quot;. */
export type PageComponentsPageComponentsBannerHomepage = AcfFieldGroup & AcfFieldGroupFields & PageComponentsPageComponentsBannerHomepage_Fields & {
  __typename?: 'PageComponentsPageComponentsBannerHomepage';
  /** Field of the &quot;group&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsBannerHomepage&quot; Field Group */
  button?: Maybe<PageComponentCloneFieldsBannerHomepageButton>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsBannerHomepage&quot; Field Group */
  description?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;image&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsBannerHomepage&quot; Field Group */
  image?: Maybe<AcfMediaItemConnectionEdge>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsBannerHomepage&quot; Field Group */
  title?: Maybe<Scalars['String']['output']>;
};

/** The &quot;PageComponentsPageComponentsBannerHomepageLayout&quot; Field Group. Added to the Schema by &quot;WPGraphQL for ACF&quot;. */
export type PageComponentsPageComponentsBannerHomepageLayout = AcfFieldGroup & AcfFieldGroupFields & PageComponentsPageComponentsBannerHomepageLayout_Fields & PageComponentsPageComponents_Layout & {
  __typename?: 'PageComponentsPageComponentsBannerHomepageLayout';
  /** Field of the &quot;group&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsBannerHomepageLayout&quot; Field Group */
  bannerHomepage?: Maybe<PageComponentsPageComponentsBannerHomepage>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};

/** Interface representing fields of the ACF &quot;PageComponentsPageComponentsBannerHomepageLayout&quot; Field Group */
export type PageComponentsPageComponentsBannerHomepageLayout_Fields = {
  /** Field of the &quot;group&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsBannerHomepageLayout&quot; Field Group */
  bannerHomepage?: Maybe<PageComponentsPageComponentsBannerHomepage>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type PageComponentsPageComponentsLesson = AcfFieldGroup & AcfFieldGroupFields & PageComponentsPageComponentsLesson_Fields & {
  __typename?: 'PageComponentsPageComponentsLesson';
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsLesson&quot; Field Group */
  description?: Maybe<Scalars['String']['output']>;
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;text&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsLesson&quot; Field Group */
  title?: Maybe<Scalars['String']['output']>;
};

/** The &quot;PageComponentsPageComponentsLessonSectionLayout&quot; Field Group. Added to the Schema by &quot;WPGraphQL for ACF&quot;. */
export type PageComponentsPageComponentsLessonSectionLayout = AcfFieldGroup & AcfFieldGroupFields & PageComponentsPageComponentsLessonSectionLayout_Fields & PageComponentsPageComponents_Layout & {
  __typename?: 'PageComponentsPageComponentsLessonSectionLayout';
  /**
   * The name of the field group
   * @deprecated Use __typename instead
   */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
  /** Field of the &quot;group&quot; Field Type added to the schema as part of the &quot;PageComponentsPageComponentsLessonSectionLayout&quot; Field Group */
  lesson?: Maybe<PageComponentsPageComponentsLesson>;
};
// ...
export type PageComponentsPageComponents_Layout = {
  /** The name of the ACF Flex Field Layout */
  fieldGroupName?: Maybe<Scalars['String']['output']>;
};
// ...
export type RootQuery = {
// ...
  lesson?: Maybe<Lesson>;
// ...
  lessons?: Maybe<RootQueryToLessonConnection>;
// ...
  page?: Maybe<Page>;
// ...
};
// ...
export type RootQueryToLessonConnection = Connection & LessonConnection & {
  __typename?: 'RootQueryToLessonConnection';
  /** Edges for the RootQueryToLessonConnection connection */
  edges: Array<RootQueryToLessonConnectionEdge>;
  /** The nodes of the connection, without the edges */
  nodes: Array<Lesson>;
  /** Information about pagination in a connection. */
  pageInfo: RootQueryToLessonConnectionPageInfo;
};
// ...
export type Template_PageNoTitle = ContentTemplate & {
  __typename?: 'Template_PageNoTitle';
  /** The name of the template */
  templateName?: Maybe<Scalars['String']['output']>;
};
// ...
export type LessonFragment = { __typename?: 'RootQuery', lessons?: { __typename?: 'RootQueryToLessonConnection', nodes: Array<{ __typename?: 'Lesson', title?: string | null, slug?: string | null, date?: string | null, content?: string | null, contentType?: { __typename?: 'ContentNodeToContentTypeConnectionEdge', node: { __typename?: 'ContentType', description?: string | null } } | null, featuredImage?: { __typename?: 'NodeWithFeaturedImageToMediaItemConnectionEdge', node: { __typename?: 'MediaItem', mediaItemUrl?: string | null } } | null, lessonCategories?: { __typename?: 'LessonToLessonCategoryConnection', nodes: Array<{ __typename?: 'LessonCategory', name?: string | null }> } | null }> } | null };
// ...
export type GetPageQuery = { __typename?: 'RootQuery', page?: { __typename?: 'Page', title?: string | null, uri?: string | null, template?: { __typename: 'DefaultTemplate', templateName?: string | null } | { __typename?: 'Template_PageNoTitle' } | null, pageComponents?: { __typename?: 'PageComponents', pageComponents?: Array<{ __typename: 'PageComponentsPageComponentsBannerHomepageLayout', bannerHomepage?: { __typename?: 'PageComponentsPageComponentsBannerHomepage', title?: string | null, description?: string | null, image?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', mediaItemUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, button?: { __typename?: 'PageComponentCloneFieldsBannerHomepageButton', label?: string | null, link?: string | null } | null } | null } | { __typename: 'PageComponentsPageComponentsLessonSectionLayout', lesson?: { __typename?: 'PageComponentsPageComponentsLesson', title?: string | null, description?: string | null } | null } | null> | null } | null } | null };

export type HomePageBannerFragment = { __typename?: 'PageComponentsPageComponentsBannerHomepageLayout', bannerHomepage?: { __typename?: 'PageComponentsPageComponentsBannerHomepage', title?: string | null, description?: string | null, image?: { __typename?: 'AcfMediaItemConnectionEdge', node: { __typename?: 'MediaItem', mediaItemUrl?: string | null, altText?: string | null, mediaDetails?: { __typename?: 'MediaDetails', width?: number | null, height?: number | null } | null } } | null, button?: { __typename?: 'PageComponentCloneFieldsBannerHomepageButton', label?: string | null, link?: string | null } | null } | null };

export type LessonSectionFragment = { __typename?: 'PageComponentsPageComponentsLessonSectionLayout', lesson?: { __typename?: 'PageComponentsPageComponentsLesson', title?: string | null, description?: string | null } | null };